Share bridge event kind resolver and pushUniqueLine across channels
Lark and weixin runtime-state modules each carried their own copy of
the bridge-event-kind resolver and the dedupe-and-cap helper. Move
both into channel/shared/bridge-event-helpers.ts; each module keeps a
thin wrapper so per-platform caps and call-site readability stay.
